Moved constants and handler packages to internal in the atlona binding
