+++ addons/binding/org.openhab.binding.atlona/META-INF/MANIFEST.MF
@@ -8,9 +8,6 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-SymbolicName: org.openhab.binding.atlona;singleton:=true
 Bundle-Vendor: openHAB
 Bundle-Version: 2.4.0.qualifier
-Export-Package: 
- org.openhab.binding.atlona,
- org.openhab.binding.atlona.handler
 Import-Package: 
  com.google.common.collect,
  org.apache.commons.lang,
@@ -23,8 +20,6 @@ Import-Package:
  org.eclipse.smarthome.core.thing.binding.builder,
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
- org.openhab.binding.atlona,
- org.openhab.binding.atlona.handler,
  org.osgi.framework,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
