+++ addons/binding/org.openhab.binding.atlona/src/main/java/org/openhab/binding/atlona/internal/pro3/AtlonaPro3Handler.java
@@ -24,9 +24,9 @@ import org.eclipse.smarthome.core.thing.ThingStatusDetail;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
 import org.eclipse.smarthome.core.types.State;
-import org.openhab.binding.atlona.handler.AtlonaHandler;
 import org.openhab.binding.atlona.internal.AtlonaHandlerCallback;
 import org.openhab.binding.atlona.internal.StatefulHandlerCallback;
+import org.openhab.binding.atlona.internal.handler.AtlonaHandler;
 import org.openhab.binding.atlona.internal.net.SocketChannelSession;
 import org.openhab.binding.atlona.internal.net.SocketSession;
 import org.slf4j.Logger;
